Replace usages of long with int32_t / uint32_t. Change power() and energy() functions to return int32_t instead of float, they don't support decimal values

+++ PZEM004T.cpp
@@ -46,7 +46,7 @@ PZEM004T::~PZEM004T()
         delete this->serial;
 }
 
-void PZEM004T::setReadTimeout(unsigned long msec)
+void PZEM004T::setReadTimeout(uint32_t msec)
 {
     _readTimeOut = msec;
 }
@@ -73,7 +73,7 @@ float PZEM004T::current(const IPAddress &addr)
     return (data[0] << 8) + data[1] + (data[2] / 100.0);
 }
 
-float PZEM004T::power(const IPAddress &addr)
+int32_t PZEM004T::power(const IPAddress &addr)
 {
     uint8_t data[RESPONSE_DATA_SIZE];
 
@@ -84,7 +84,7 @@ float PZEM004T::power(const IPAddress &addr)
     return (data[0] << 8) + data[1];
 }
 
-float PZEM004T::energy(const IPAddress &addr)
+int32_t PZEM004T::energy(const IPAddress &addr)
 {
     uint8_t data[RESPONSE_DATA_SIZE];
 
@@ -134,7 +134,7 @@ bool PZEM004T::recieve(uint8_t resp, uint8_t *data)
         ((SoftwareSerial *)serial)->listen();
 #endif
 
-    unsigned long startTime = millis();
+    uint32_t startTime = millis();
     uint8_t len = 0;
     while((len < RESPONSE_SIZE) && (millis() - startTime < _readTimeOut))
     {
